requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
httpx==0.25.2
pandas==2.1.3
pydantic==2.5.2
plaid-python==18.0.0 